Move showWaitCursorWhileWaiting() from __init__
