+++ src/ert/gui/ertwidgets/__init__.py
@@ -1,10 +1,3 @@
-from collections.abc import Callable
-from typing import Any
-
-from PyQt6.QtCore import Qt
-from PyQt6.QtGui import QCursor
-from PyQt6.QtWidgets import QApplication
-
 from .analysismoduleedit import AnalysisModuleEdit
 from .checklist import CheckList
 from .closabledialog import ClosableDialog
@@ -27,20 +20,7 @@ from .searchbox import SearchBox
 from .stringbox import StringBox
 from .suggestor import Suggestor
 from .textbox import TextBox
-
-
-def showWaitCursorWhileWaiting(func: Callable[..., Any]) -> Callable[..., Any]:
-    """A function decorator to show the wait cursor while the function is working."""
-
-    def wrapper(*arg: Any) -> Any:
-        QApplication.setOverrideCursor(QCursor(Qt.CursorShape.WaitCursor))
-        try:
-            return func(*arg)
-        finally:
-            QApplication.restoreOverrideCursor()
-
-    return wrapper
-
+from .utils import showWaitCursorWhileWaiting
 
 __all__ = [
     "ActiveRealizationsModel",
